utils/image_diagnosis.py
from PIL import Image
import io

def diagnose_image(image_file) -> str:
    """
    Dummy function to diagnose skin rash images.
    Replace with actual ML model logic as needed.
    """
    try:
        image = Image.open(image_file)
        # Demo: Just return a placeholder diagnosis
        return "Possible diagnosis: Mild dermatitis. Please consult a dermatologist for confirmation."
    except Exception as e:
        return f"Error processing image: {e}"
