requirements.txt
streamlit
pillow
transformers
langdetect
speechrecognition
torch
